Added new rapid fire bullet type

+++ core/src/main/java/com/sixmoney/gigagal/entities/Bullet.java
@@ -14,7 +14,7 @@ public class Bullet {
     protected Direction direction;
     protected Vector2 position;
     protected Level level;
-    protected int damage;
+    protected float damage;
 
     public boolean active;
